+++ hd-progress/element.php
@@ -15,8 +15,8 @@ return [
 
 			$metadata->set('script:builder-hd-progress', ['src' => Path::get('./js/hd-progress.js'), 'defer' => true]);
 
-            // Don't render element if content fields are empty
-            return Str::length($node->props['content']) || ($node->props['max'] && $node->props['stop'] && $node->props['animation_step'] && $node->props['animation_speed']);
+			// Don't render element if content fields are empty
+			return Str::length($node->props['content']) || ($node->props['max'] && $node->props['stop'] && $node->props['animation_step'] && $node->props['animation_speed']);
 
 		},
 
@@ -24,7 +24,14 @@ return [
 
 	'updates' => [
 
-		//
+		'2.4.12' => function ($node) {
+
+			if (isset($node->props['value'])) {
+				$node->props['start'] = $node->props['value'];
+				unset($node->props['value']);
+			}
+
+		},
 
 	],